use *Reviver to restore object from localStorage (used only in devxpress demo)
