+++ assets/js/devxpress-angular/package.json
@@ -21,6 +21,7 @@
     "@angular/platform-browser": "^5.2.0",
     "@angular/platform-browser-dynamic": "^5.2.0",
     "@angular/router": "^5.2.0",
+    "@rebolon/json-reviver": "0.0.4",
     "core-js": "^2.4.1",
     "devextreme": "^17.2.6",
     "devextreme-angular": "^17.2.6",
