+++ assets/js/devxpress-angular/src/reviver/library/bookReviver.ts
@@ -1,9 +1,9 @@
-import {Book} from "../../../entities/library/book"
-import {ItemAbstractReviver} from "../itemAbstractReviver";
+import {Book} from "../../entities/library/book"
+import {ItemAbstractReviver} from "@rebolon/json-reviver/src";
 import {AuthorsReviver} from "./authorsReviver";
 import {SerieReviver} from "./serieReviver";
 import {EditorsReviver} from "./editorsReviver";
-import {accessor} from "../../accessor";
+import {Accessor} from "@rebolon/json-reviver/src/accessor";
 
 export class BookReviver extends ItemAbstractReviver
 {
@@ -81,17 +81,17 @@ export class BookReviver extends ItemAbstractReviver
         // that's why i don't add reviews here
         return {
             'authors': {
-                'converter': this.authorsReviver,
+                'reviver': this.authorsReviver,
                 'setter': 'addAuthor',
                 'cb': function (relation, entity) {
-                    accessor('book', relation, entity)
+                    Accessor('book', relation, entity)
                 },
             },
             'editors': {
-                'converter': this.editorsReviver,
+                'reviver': this.editorsReviver,
                 'setter': 'addEdition',
                 'cb': function (relation, entity) {
-                    accessor('book', relation, entity)
+                    Accessor('book', relation, entity)
                 },
             },
         }
@@ -106,7 +106,7 @@ export class BookReviver extends ItemAbstractReviver
     {
         return {
             'serie': {
-                'converter': this.serieReviver,
+                'reviver': this.serieReviver,
                 'registryKey': 'serie',
             },
         }
