--- assets/js/devxpress-angular/src/tools/reviver/library/entityInterface.ts
@@ -1,3 +1,0 @@
-export interface EntityInterface
-{
-}
